add tree bricks, track type/origin checkboxes, etc
untested. a lot of the inner workings of how maps load were changed to
account for the now flexible map type and origin settings

this commit also adds super's track edits which use the tree bricks, and
a few more tracks edited by super or sent to me by them.